Add GetAdaptersAddress()-based interface enumeration to libisc Windows code,
...runtime linked on Windows XP SP1 and later only, which provides
correct interface name with each address.

+++ lib/isc/win32/include/isc/once.h
@@ -27,7 +27,7 @@ ISC_LANG_BEGINDECLS
 
 typedef struct {
 	int status;
-	int counter;
+	LONG counter;
 } isc_once_t;
 
 #define ISC_ONCE_INIT_NEEDED 0
